+++ src/layouts/stack/component.tsx
@@ -69,6 +69,9 @@ const StyledStack = styled.div<IStyledStackProps>`
   &.scrollableHorizontally {
     overflow-x: auto;
   }
+  &.wrapItems {
+    flex-wrap: wrap;
+  }
 `;
 
 interface IStackProps extends IMultiAnyChildProps, IPaddingViewPaddingProps {
@@ -89,6 +92,7 @@ interface IStackProps extends IMultiAnyChildProps, IPaddingViewPaddingProps {
   childAlignmentResponsive?: ResponsiveField<Alignment>;
   contentAlignment: Alignment;
   contentAlignmentResponsive?: ResponsiveField<Alignment>;
+  shouldWrapItems?: boolean;
 }
 
 export const Stack = (props: IStackProps): React.ReactElement => {
@@ -103,11 +107,12 @@ export const Stack = (props: IStackProps): React.ReactElement => {
 
   const defaultGutter = props.defaultGutter || PaddingSize.Default;
   const shouldAddGutters = props.shouldAddGutters && defaultGutter !== PaddingSize.None;
+  const shouldWrapItems = props.shouldWrapItems || false;
   return (
     <PaddingView paddingTop={paddingTop} paddingBottom={paddingBottom} paddingRight={paddingRight} paddingLeft={paddingLeft} {...props as IPaddingViewPaddingProps}>
       <StyledStack
         id={props.id}
-        className={getClassName(Stack.displayName, props.isScrollableVertically && 'scrollableVertically', props.isScrollableHorizontally && 'scrollableHorizontally')}
+        className={getClassName(Stack.displayName, props.isScrollableVertically && 'scrollableVertically', props.isScrollableHorizontally && 'scrollableHorizontally', shouldWrapItems && 'wrapItems')}
         theme={theme}
         $direction={{ base: props.direction, ...props.directionResponsive }}
         childAlignment={{ base: props.childAlignment, ...props.childAlignmentResponsive }}