+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MINOR] Updated `LayerContainer` wrapper component
 - [MINOR] Correct scrollability in `Stack`
 - [MINOR] Updated to `Portal` to have offset and placement props
+- [MINOR] Updated `Stack` to have `shouldWrapItems` prop to wrap items when it overflows
+- [MINOR] Updated `Box` props so that children are optional
 
 ### Removed